requirements.txt
requests
pandas==2.1.4
sqlalchemy==1.4.20
psycopg2-binary
python-dotenv
sshtunnel
Pillow
mage-ai[psycopg2]==0.9.76
tornado==6.4.2
jupyter-client==7.4.4
